[TASK] Re-add docblocks to all classes and interfaces
This is need to work around a PHP bug resulting in a (random?)
docblock being returned when calling getDocComment() on a class
reflection of a class that has no docblock.

See PHP bug 55156. Although reported to be fixed as of 5.3.7 the
test for the bug fails in 5.3.8 (for 5.4.0 the test passes).

The bug then causes errors to be reported by Doctrine's annotation
reader, because it "sees" annotations on classes that don't have
them, and thus lack needed use statements. Bummer.

Note: The docblocks have been added before but got lost with the
latest update of the package.

Change-Id: Ic478ecbebf893eb2a0d1dffa9688179acffb782e
Related: #27627