[!!!][BUGFIX] Adapt to changes in imagine master
Imagine is now identified as "imagine.imagine" instead of
"imagine.Imagine" so object configuration has to be adapted.
Additionally the version requirement is raised to minimum 0.5
to avoid wrong object configurations with older imagine versions.

To fully fix the problem you need to delete your
Configuration/PackageStates.php file as imagine is registered
with both package keys after an update and so will stil throw
the same error. That is why this change is breaking. The change
itself is not breaking but the problem it tries to fix needs
manual deletion of the PackageStates.

Change-Id: If496b91861828e2e269e8187d0fcfa69ac3327aa