[TASK] Add Settings for custom ICC profile
Imagine 0.6 support ICC profile, but use internal the constants __DIR__
to load the profile. This change add an aspect to allow the ICC profile
configuration in Settings.yaml and avoid critical error in proxy class.

The new settings allow to change the default ICC profile. The ICC
profile can be stored in custom package if needed.

This change also add some documentation and schema validation for the
package settings.

Change-Id: Ia38f6d12b13d2b4e18deac9bcedb2c038d4ea206
Releases: master